moved entities to common module for modularity
