+++ apps/klubiq-dashboard/src/users/users.module.ts
@@ -4,10 +4,6 @@ import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { DatabaseModule } from '@app/common';
 import { OrganizationUser } from './entities/organization-user.entity';
-import { OrganizationRole } from './entities/organization-role.entity';
-import { Feature } from './entities/feature.entity';
-import { FeaturePermission } from './entities/feature-permission.entity';
-import { Permission } from './entities/permission.entity';
 import { AuthModule } from '@app/auth';
 
 @Module({
@@ -16,10 +12,6 @@ import { AuthModule } from '@app/auth';
 		AuthModule,
 		TypeOrmModule.forFeature([
 			OrganizationUser,
-			OrganizationRole,
-			Feature,
-			FeaturePermission,
-			Permission,
 		]),
 	],
 	controllers: [UsersController],
